requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.7.1
pydantic-settings>=2.2.1
PyJWT>=2.8.0
python-dotenv>=1.0.1
aiofiles>=23.0.0
